+++ lib/src/presentation/ui/templates/ride_widgets/ride_detail_widgets.dart
@@ -259,7 +259,7 @@ Widget rideDetailsInProgressAndFinishedWidget(
   );
 }
 
-Widget rideRatingWidget(BuildContext context, double fare,String drivername) {
+Widget rideRatingWidget(BuildContext context, int fare,String drivername) {
   return Column(
     children: [
       Container(
